layered: Model order comparator cleanup.
